refactor: tree: Move ReferenceCounted to util (#14353)
## Description

Move ReferenceCounted to util where it can be used by AnchorSet.
