+++ packages/dds/tree/src/feature-libraries/chunked-forest/uniformChunk.ts
@@ -14,9 +14,9 @@ import {
 	Value,
 	PathRootPrefix,
 } from "../../core";
-import { compareArrays, fail } from "../../util";
+import { compareArrays, fail, ReferenceCountedBase } from "../../util";
 import { prefixFieldPath, prefixPath, SynchronousCursor } from "../treeCursorUtils";
-import { ChunkedCursor, cursorChunk, dummyRoot, ReferenceCountedBase, TreeChunk } from "./chunk";
+import { ChunkedCursor, cursorChunk, dummyRoot, TreeChunk } from "./chunk";
 
 /**
  * Create a tree chunk with ref count 1.
