+++ packages/dds/tree/src/util/index.ts
@@ -62,3 +62,4 @@ export {
 	RecursiveReadonly,
 	zipIterables,
 } from "./utils";
+export { ReferenceCountedBase, ReferenceCounted } from "./referenceCounting";
